Fix narrowing and oveeride warnings for Clang
Fhicl-cpp types submodule emits two main warnings for Clang:

- Narrowing warnings for size_t non-type template parameters
- Missing override on concrete class virtual functions

Fix narrowing warnings by explicit static_cast to size_t. This
should have the exatc same behaviour.

Mark overidden base class virtual functions explicitly as override.

+++ fhiclcpp/types/OptionalSequence.h
@@ -23,7 +23,7 @@ namespace fhicl {
   //==================================================================
   // e.g. OptionalSequence<int,4> ====> std::array<int,4>
   //
-  template<typename T, std::size_t N = -1>
+  template<typename T, std::size_t N = static_cast<std::size_t>(-1)>
   class OptionalSequence final :
     public  detail::SequenceBase,
     private detail::RegisterIfTableMember {
@@ -79,7 +79,7 @@ namespace fhicl {
   // e.g. OptionalSequence<int> ====> std::vector<int>
   //
   template<typename T>
-  class OptionalSequence<T,-1> final :
+  class OptionalSequence<T, static_cast<std::size_t>(-1)> final :
     public  detail::SeqVectorBase,
     private detail::RegisterIfTableMember {
   public:
